Add files via upload
The users table can have 0 to many people (A user can have 0 or more people in their contacts). And the People table can have a 1 to many users (a contact has to have at least 1 person, but can be in multiple user's contacts, however in our database it'll show up as a duplicate person with a different ID).